refactor: move App too src, kill for good @cow-react
